backend/requirements.txt
# ==============================================================================
# AI Mock Interview System - Backend Dependencies
# ==============================================================================

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI/ML Libraries
openai==1.10.0
google-generativeai==0.3.2
transformers==4.37.0
torch==2.1.2

# Speech-to-Text
openai-whisper==20231117
speechrecognition==3.10.1
pydub==0.25.1

# NLP Evaluation
nltk==3.8.1
language-tool-python==2.7.1
sentence-transformers==2.3.1
numpy==1.26.3
pandas==2.1.4

# File Handling
aiofiles==23.2.1
python-magic==0.4.27

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
uuid==1.30

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==23.12.1
isort==5.13.2
